+++ packages/implementations/react/src/atomic/Toast/src/context/Manager/Manager.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { ContextProps } from "./manager.types";
+
+export const Context = createContext<ContextProps>({} as ContextProps);