+++ packages/implementations/react/src/atomic/Toast/src/index.ts
@@ -1,5 +1,6 @@
 import { Toast } from "./Toast";
 
 export { Toast } from "./Toast";
+export { useToast } from "./hooks";
 export type { ToastProps } from "./toast.types";
 export default Toast;